# Switch from gob to jsoniter for serialization