--- a/data/token.go
+++ b/data/token.go
@@ -16,20 +16,19 @@
 package data
 
 import (
-	"bytes"
 	"encoding/base64"
-	"encoding/gob"
 	"errors"
 	"time"
 
 	"github.com/cybercryptio/d1-lib/v2/crypto"
+	json "github.com/json-iterator/go"
 )
 
 // Validity period of tokens created with D1.CreateToken.
 const TokenValidity = time.Hour
 
 // Token contains arbitrary data along with an expiry time. Note: All fields need to be exported in
-// order for gob to serialize them.
+// order to serialize them.
 type Token struct {
 	// Arbitrary plaintext data.
 	Plaintext []byte
@@ -56,7 +55,7 @@ func NewToken(plaintext []byte, validityPeriod time.Duration) Token {
 
 // Seal encrypts the token.
 func (t *Token) Seal(cryptor crypto.CryptorInterface) (SealedToken, error) {
-	associatedData, err := t.ExpiryTime.GobEncode()
+	associatedData, err := t.ExpiryTime.MarshalBinary()
 	if err != nil {
 		return SealedToken{}, err
 	}
@@ -84,7 +83,7 @@ func (t *SealedToken) Unseal(cryptor crypto.CryptorInterface) (Token, error) {
 		return Token{}, errors.New("Token expired")
 	}
 
-	associatedData, err := t.ExpiryTime.GobEncode()
+	associatedData, err := t.ExpiryTime.MarshalBinary()
 	if err != nil {
 		return Token{}, err
 	}
@@ -100,13 +99,12 @@ func (t *SealedToken) Unseal(cryptor crypto.CryptorInterface) (Token, error) {
 
 // String serializes the sealed token into a raw base 64 URL encoded format.
 func (t *SealedToken) String() (string, error) {
-	var buffer bytes.Buffer
-	enc := gob.NewEncoder(&buffer)
-	if err := enc.Encode(t); err != nil {
+	b, err := json.Marshal(t)
+	if err != nil {
 		return "", err
 	}
 
-	return base64.RawURLEncoding.EncodeToString(buffer.Bytes()), nil
+	return base64.RawURLEncoding.EncodeToString(b), nil
 }
 
 // TokenFromString takes a raw base 64 URL encoded token and deserializes it.
@@ -116,8 +114,7 @@ func TokenFromString(tokenString string) (SealedToken, error) {
 		return SealedToken{}, err
 	}
 	var token SealedToken
-	dec := gob.NewDecoder(bytes.NewReader(tokenBytes))
-	if err := dec.Decode(&token); err != nil {
+	if err := json.Unmarshal(tokenBytes, &token); err != nil {
 		return SealedToken{}, err
 	}
 	return token, nil