### Description
Did some profiling and found that most of our time was spent calling `gob.Encode`/`gob.Decode`. I had a look at the benchmarks [here](https://github.com/alecthomas/go_serialization_benchmarks) and [here](https://github.com/smallnest/gosercomp) to see if I could find something better. 

The best performance is clearly achieved with serializers that generate dedicated code ahead of time. Sadly, most of those do not support types from imported packages. 

I settled on using [jsoniter](https://github.com/json-iterator/go) which is significantly faster then `gob` and has less overhead. The speedup can be seen in the following benchmark:

```
name       old time/op  new time/op  delta
Encrypt-8   171µs ±25%    90µs ±26%  -47.41%  (p=0.000 n=100+97)
Decrypt-8   135µs ±20%    68µs ±24%  -49.59%  (p=0.000 n=96+99)
```
Approximately 50% improvement on simple encrypt/decrypt calls. 

### Related PR
Closes DEV-437

### Type(s) of Change (Split the PR if you check many)
- [ ] Added/updated user feature
- [x] Technical change
- [ ] Refactor
- [ ] Bugfix
- [ ] Dependency update
- [ ] Added/updated tests
- [ ] Updated documentation
- [ ] Workflow/tooling changes

### Testing Checklist
- [ ] I have added/updated unit tests for functional changes.
- [ ] I have added/updated integration tests for changes that affect dependent systems.
- [ ] I have added/updated end-to-end tests for feature changes.

### General Checklist
- [x] I have pulled in the latest changes from master.
- [ ] I have added the license header to new source code files.
- [x] I have checked the code with linting tools.
- [x] I have successfully run all tests.
- [ ] I have updated relevant CI/CD workflows.
- [ ] I have updated/added relevant (internal and external) documentation (readme, doc comments, etc).
- [ ] I have updated the dependency map to reflect my changes.
- [x] I have spent some time looking over the full diff before creating this PR.