@philip-cc 2022-09-15
> Will this break support for already existing setups?

Yes, but we're already bumping to 2.0, so I thought now was a good time to do it.